# optimize embedding bag